// projects/codx-ac/src/lib/models/DeductPrepaidExpenses.model.ts
export class DeductPrepaidExpenses{
    recID: any;
    runType: any;
    runDate: any;
    runMode: any;
    periodID: any;
    fromDate: any;
    toDate: any;
    assetGroupID: any;
    memo: any;
    memo2: any;
    status: any;
    paras: any;
    completed: any;
    deprModelID: any;
    errorLog: any;
    autoCreated: any;
    owner: any;
    buid: any;
    createdOn: any;
    createdBy: any;
    modifiedOn: any;
    modifiedBy: any;
    assetID: any;
    calcGroupID: any;
}